+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - Fixed himg require from native extension
+- Catch rust panics and wrap with ruby Error
 
 ## [0.0.9] - 2025-07-31